generator/requirements.txt
google-genai
fastapi[standard]